+++ client/src/components/Update.jsx
@@ -1,4 +1,5 @@
 import { useState } from 'react'
+import { useNavigate } from 'react-router-dom'
 import Button from '@mui/material/Button'
 import TextField from '@mui/material/TextField'
 import Dialog from '@mui/material/Dialog'
@@ -7,10 +8,25 @@ import DialogContent from '@mui/material/DialogContent'
 import DialogTitle from '@mui/material/DialogTitle'
 import MenuItem from '@mui/material/MenuItem';
 import axios from 'axios'
-import { EditablePages, EditableTitles, EditableSelects } from '../configs/pages'
+import { EditablePages, AdminEditableTitles, UserEditableTitles, EditableSelects } from '../configs/pages'
+import { getToday } from '../utils/DateUtil'
 import { options } from '../configs/options'
+import { getUser } from '../storage/userSlice.js'
+
+const getEditableTitles = (user) => {
+    let editableTitles = AdminEditableTitles
+    if (!user.IsAdmin) {
+        editableTitles = UserEditableTitles
+    }
+    return editableTitles
+}
 
 const Update = ({page, columns, data, setData, open, setOpen, rowData}) => {
+
+    const user = getUser()
+    const editableTitles = getEditableTitles(user)
+    const previousStatus = rowData.status
+    const navigate = useNavigate()
 
     const [focus, setFocus] = useState('info')
     const [value, setValue] = useState(rowData)
@@ -28,8 +44,33 @@ const Update = ({page, columns, data, setData, open, setOpen, rowData}) => {
         const url = '/api/' + page + '/update'
         if (EditablePages.includes(page)) {
             try {
-                await axios.post(url, value)
-                updateData()
+                if (page === 'approval') {
+                    if (previousStatus === value.status) { 
+                        alert('status가 바뀌지 않았습니다.')
+                        navigate('/approvalhistory')
+                    } else if (user.isAdmin) {
+                        if (previousStatus === 'Pending' || (previousStatus === 'Active' && value.status === 'Cancel')) {
+                            const res = await axios.post(url, value)
+                            setValue(res.data)
+                            updateData()
+                        } else {
+                            alert('변경할 수 없는 조건입니다.')
+                            navigate('/approvalhistory')
+                        }     
+                    } else if ((previousStatus === 'Pending' && value.status === 'Cancel') ||
+                        (previousStatus === 'Active' && value.status === 'Cancel' && value.start) > getToday()) {
+                        const res = await axios.post(url, value)
+                        setValue(res.data)
+                        updateData()
+                    } else {
+                        alert('변경할 수 없는 조건입니다.')
+                        navigate('/approvalhistory')
+                    }
+                } else {
+                    const res = await axios.post(url, value)
+                    setValue(res.data)
+                    updateData()
+                }
             } catch (err) {
                 console.log(url, err)
             }
@@ -60,9 +101,10 @@ const Update = ({page, columns, data, setData, open, setOpen, rowData}) => {
                                 label={item.accessor}
                                 fullWidth
                                 variant='outlined'
-                                defaultValue={value[item.accessor]}
+                                defaultValue={value[item.accessor]?value[item.accessor]:''}
                                 key={index}
                                 onChange={handleChange}
+                                autoComplete='false'
                             >  
                                 {options[item.accessor].map((option) => (
                                     <MenuItem
@@ -73,7 +115,7 @@ const Update = ({page, columns, data, setData, open, setOpen, rowData}) => {
                                     </MenuItem>
                                 ))}
                             </TextField>
-                        ):EditableTitles.includes(item.accessor)?(
+                        ):editableTitles.includes(item.accessor)?(
                             <TextField
                                 autoFocus={focus===item.accessor}
                                 margin='dense'