+++ client/src/configs/pages.js
@@ -1,4 +1,5 @@
 export const UpdatePages = ['board', 'report']
 export const EditablePages = ['location', 'device', 'employee', 'approval']
-export const EditableTitles = ['info', 'type', 'location', 'beginDate', 'charge', 'email', 'department', 'rank', 'position', 'regular', 'mode', 'attendMode', 'latitude', 'longitude', 'dev', 'reason', 'etc', 'status']
-export const EditableSelects = ['regular', 'mode', 'attendMode', 'reason', 'status']
+export const AdminEditableTitles = ['info', 'type', 'location', 'charge', 'beginDate', 'email', 'department', 'rank', 'position', 'regular', 'mode', 'attendMode', 'latitude', 'longitude', 'dev', 'status']
+export const UserEditableTitles = ['info', 'type', 'location', 'charge', 'status']
+export const EditableSelects = ['regular', 'mode', 'attendMode', 'status']